+++ R/04_phaseII.R
@@ -119,6 +119,7 @@
 #' Functional Regression Control Chart.
 #' \emph{Technometrics}. <doi:10.1080/00401706.2020.1753581>
 #'
+#' @examples
 #' library(funcharts)
 #' data("air")
 #' air <- lapply(air, function(x) x[1:220, , drop = FALSE])